requirements.txt
Django==1.10.5
git+https://github.com/parksandwildlife/dpaw-utils.git@0.3a15#egg=dpaw-utils
django-model-utils==2.6.1
webtemplate-dpaw==0.4.5
